requirements.txt
streamlit>=1.28.0
praw>=7.7.1
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
nltk>=3.8.1
requests>=2.31.0
sqlalchemy>=2.0.0
pytest>=7.4.0
scikit-learn>=1.3.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain
langchain-community